+++ source/renderer/app/api/utils/patchAdaApi.js
@@ -9,6 +9,11 @@ import { getLatestAppVersion } from '../nodes/requests/getLatestAppVersion';
 import { GenericApiError } from '../common/errors';
 import { Logger } from '../../utils/logging';
 import packageJson from '../../../../../package.json';
+
+// domains
+import Wallet from '../../domains/Wallet';
+import StakePool from '../../domains/StakePool';
+
 import type {
   GetNetworkInfoResponse,
   NetworkInfoResponse,
@@ -19,9 +24,6 @@ import type {
 } from '../nodes/types';
 import type { GetNewsResponse } from '../news/types';
 import { EPOCH_LENGTH_ITN } from '../../config/epochsConfig';
-
-// domains
-import Wallet from '../../domains/Wallet';
 
 let LATEST_APP_VERSION = null;
 let SYNC_PROGRESS = null;
@@ -146,14 +148,14 @@ export default (api: AdaApi) => {
     APPLICATION_VERSION = applicationVersion;
   };
 
-  api.setFakeNewsFeedJsonForTesting = (fakeNewsfeedJson: ?GetNewsResponse) => {
+  api.setTestingNewsFeed = (testingNewsFeedData: ?GetNewsResponse) => {
     const { version: packageJsonVersion } = packageJson;
-    if (!fakeNewsfeedJson) {
+    if (!testingNewsFeedData) {
       FAKE_NEWSFEED_JSON = null;
       return;
     }
     // Always mutate newsfeed target version to current app version
-    const newsFeedItems = map(fakeNewsfeedJson.items, item => {
+    const newsFeedItems = map(testingNewsFeedData.items, item => {
       return {
         ...item,
         target: {
@@ -166,7 +168,7 @@ export default (api: AdaApi) => {
     });
 
     FAKE_NEWSFEED_JSON = {
-      ...fakeNewsfeedJson,
+      ...testingNewsFeedData,
       items: newsFeedItems,
     };
   };
@@ -198,6 +200,18 @@ export default (api: AdaApi) => {
     const wallet = getModifiedWallet();
     api.getWallets = async (): Promise<Array<Wallet>> => [wallet];
     return wallet;
+  };
+
+  api.setTestingWallets = (testingWalletsData: Array<Object>): void => {
+    api.getWallets = (): Array<Wallet> =>
+      testingWalletsData.map((wallet: Object) => new Wallet(wallet));
+  };
+
+  api.setTestingStakePools = (testingStakePoolsData: Array<Object>): void => {
+    api.getStakePools = (): Array<StakePool> =>
+      testingStakePoolsData.map(
+        (stakePool: Object) => new StakePool(stakePool)
+      );
   };
 
   api.resetTestOverrides = () => {